% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/already_vectorized.R
\name{already_vectorized}
\alias{already_vectorized}
\title{Define the function F of a glass of wine by using only vectorized operations}
\usage{
already_vectorized(x)
}
\arguments{
\item{x}{the variable of our function F}
}
\value{
the value of the function F evaluated for X
}
\description{
Define the function F of a glass of wine by using only vectorized operations
}
\examples{
already_vectorized(3)
}
