% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark1_results.R
\name{benchmark1_results}
\alias{benchmark1_results}
\title{Extract the results of the Benchmarks including the vectorized function}
\usage{
benchmark1_results(i)
}
\arguments{
\item{i}{X values to use when evaluating the functions to benchmark}
}
\value{
List of benchmarks
}
\description{
Extract the results of the Benchmarks including the vectorized function
}
