% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_results.R
\name{benchmark_results}
\alias{benchmark_results}
\title{Extracting the result of our benchmarks}
\usage{
benchmark_results(i)
}
\arguments{
\item{i}{corresponds to the list of X values to pass to the F functions to be evaluated in the benchmark.}
}
\value{
List of benchmarks
}
\description{
Extracting the result of our benchmarks
}
\examples{
benchmark_results(20)
}
