% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_plots.R
\name{curve_plots}
\alias{curve_plots}
\title{Generic function to plot the results using the curve method}
\usage{
curve_plots(plot_func, x_values, graph_title)
}
\arguments{
\item{plot_func}{The function that we want to evaluate and plot (func_for, func_apply, etc...)}

\item{x_values}{The list of values for which we want to evaluate our function F}

\item{graph_title}{The title of the plot}
}
\value{
Curve plot of the evaluated function
}
\description{
Generic function to plot the results using the curve method
}
