% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_apply.R
\name{func_apply}
\alias{func_apply}
\title{Evaluating F function by using base::sapply function}
\usage{
func_apply(x)
}
\arguments{
\item{x}{a list of values for X}
}
\value{
a list of Y values of the evaluated function F
}
\description{
Evaluating F function by using base::sapply function
}
