% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_for.R
\name{func_for}
\alias{func_for}
\title{Evaluating function F using for loop}
\usage{
func_for(x)
}
\arguments{
\item{x}{a list of values of X to use when evaluating the function}
}
\value{
a list of values y corresponding to the evaluation of F for the values of X
}
\description{
Evaluating function F using for loop
}
