% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_map.R
\name{func_map}
\alias{func_map}
\title{Evaluate function F by using purrr::map function for the values of X}
\usage{
func_map(x)
}
\arguments{
\item{x}{a list of numeric values}
}
\value{
a list of numeric values for the evaluation of F
}
\description{
Evaluate function F by using purrr::map function for the values of X
}
