% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_vectorize.R
\name{func_vectorize}
\alias{func_vectorize}
\title{Evaluating the function F using Vectorize}
\usage{
func_vectorize(x)
}
\arguments{
\item{x}{a list of values of X}
}
\value{
a list of Y values of the evaluated function
}
\description{
Evaluating the function F using Vectorize
}
