% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_input.R
\name{generate_input}
\alias{generate_input}
\title{Generating the input size for the benchmarking}
\usage{
generate_input(size)
}
\arguments{
\item{size}{}
}
\value{
Input to use in the benchmark function
}
\description{
Generating the input size for the benchmarking
}
