% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass_function.R
\name{glass_function}
\alias{glass_function}
\title{Defining function F representing the cut of a glass of wine}
\usage{
glass_function(x)
}
\arguments{
\item{x}{the value of x parameter (x-axis)}
}
\value{
y value of function F evaluated for x
}
\description{
Defining function F representing the cut of a glass of wine
}
