% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integral_i.R
\name{integral_i}
\alias{integral_i}
\title{Integrate the function I}
\usage{
integral_i(lowerb, upperb)
}
\arguments{
\item{lowerb}{Lower bound of the integral}

\item{upperb}{Upper bound of the integral}
}
\value{
Evaluation of the integrate() function using already_vectorized()
}
\description{
Integrate the function I
}
