% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integral_v.R
\name{integral_v}
\alias{integral_v}
\title{Integrate the function V (Volume of the Glass)}
\usage{
integral_v(lowerb, upperb)
}
\arguments{
\item{lowerb}{Lower bound of the integral}

\item{upperb}{Upper bound of the integral}
}
\value{
The numeric value representing the Volume of the glass
}
\description{
Integrate the function V (Volume of the Glass)
}
