% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_expression_i.R
\name{math_expression_i}
\alias{math_expression_i}
\title{Computing the result of mathematical expression I}
\usage{
math_expression_i()
}
\value{
Numeric value of the result
}
\description{
Computing the result of mathematical expression I
}
