% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_expression_v.R
\name{math_expression_v}
\alias{math_expression_v}
\title{Computing the volume of the glass without using the integral}
\usage{
math_expression_v()
}
\value{
Numeric value of the volume of the glass
}
\description{
Computing the volume of the glass without using the integral
}
