% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_result.R
\name{optimize_result}
\alias{optimize_result}
\title{Use optimize function to find the upper bound of our V function to have a volume of 0.2L}
\usage{
optimize_result()
}
\value{
the upper bound
}
\description{
Use optimize function to find the upper bound of our V function to have a volume of 0.2L
}
