% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_benchmark.R
\name{plot_benchmark}
\alias{plot_benchmark}
\title{Plotting the results of the benchmarks}
\usage{
plot_benchmark(b_results, i_sizes)
}
\arguments{
\item{b_results}{List of the results of the benchmarks after using microbenchmark}

\item{i_sizes}{Input size of the benchmarks}
}
\value{
Boxplot of the benchmarks
}
\description{
Plotting the results of the benchmarks
}
