% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_benchmarks.R
\name{run_benchmarks}
\alias{run_benchmarks}
\title{Run microbenchmark::microbenchmark for all the functions}
\usage{
run_benchmarks(size)
}
\arguments{
\item{size}{is the input size generated by generate_input function}
}
\value{
list of benchmarks
}
\description{
Run microbenchmark::microbenchmark for all the functions
}
