% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_benchmarks1.R
\name{run_benchmarks1}
\alias{run_benchmarks1}
\title{Run the benchmark for all functions including the Vectorized function}
\usage{
run_benchmarks1(size)
}
\arguments{
\item{size}{input size of the benchmark function}
}
\value{
List of benchmarks
}
\description{
Run the benchmark for all functions including the Vectorized function
}
