% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Evaluate the simulation and return plots of the results}
\usage{
simulate()
}
\value{
Prints plots of the simulations
}
\description{
Evaluate the simulation and return plots of the results
}
