% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_party.R
\name{simulate_party}
\alias{simulate_party}
\title{Generate the numeric values to create simulation of a party}
\usage{
simulate_party()
}
\description{
Generate the numeric values to create simulation of a party
}
