% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_results.R
\name{simulate_results}
\alias{simulate_results}
\title{Evaluate the simulation and return plots of the results}
\usage{
simulate_results()
}
\value{
Prints plots of the simulations
}
\description{
Evaluate the simulation and return plots of the results
}
