% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniroot_result.R
\name{uniroot_result}
\alias{uniroot_result}
\title{Use uniroot to find the upper bound of the function V}
\usage{
uniroot_result()
}
\value{
The value of the upper bound
}
\description{
Use uniroot to find the upper bound of the function V
}
