% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_function.R
\name{volume_function}
\alias{volume_function}
\title{Finding the upper bound of the Volume function needed to have integrate_v(10,b) = 0.2}
\usage{
volume_function(b)
}
\arguments{
\item{b}{Upper bound to find}
}
\value{
b: the upper bound of the function by using uniroot
}
\description{
Finding the upper bound of the Volume function needed to have integrate_v(10,b) = 0.2
}
