% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_to_litres.R
\name{volume_to_litres}
\alias{volume_to_litres}
\title{Converting volume in cm3 to litres}
\usage{
volume_to_litres(volume)
}
\arguments{
\item{volume}{numeric value of the volume in cm3}
}
\value{
The numeric value of the volume in litres
}
\description{
Converting volume in cm3 to litres
}
